Share person_details component with name and dob
