+++ app/views/components/person_details.scala.html
@@ -1,0 +1,40 @@
+@*
+ * Copyright 2018 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import uk.gov.hmrc.play.views.html._
+
+@(form: Form[_])(implicit messages: Messages)
+
+@components.input_text(
+    field = form("firstName"),
+    label = messages("messages__common__firstname")
+)
+
+@components.input_text(
+    field = form("middleName"),
+    label = messages("messages__common__middlename")
+)
+
+@components.input_text(
+    field = form("lastName"),
+    label = messages("messages__common__lastname")
+)
+
+@components.input_date(
+    field = form("date"),
+    label = messages("messages__common__dob"),
+    hint = Some(messages("messages__common__dob__hint"))
+)
