+++ app/views/register/establishers/individual/establisherDetails.scala.html
@@ -35,26 +35,7 @@
 
         @components.heading(headingKey = "messages__establisher_individual__title", secondaryHeaderKey = Some(schemeName))
 
-        @components.input_text(
-            field = form("firstName"),
-            label = messages("messages__common__firstname")
-        )
-
-        @components.input_text(
-            field = form("middleName"),
-            label = messages("messages__common__middlename")
-        )
-
-        @components.input_text(
-            field = form("lastName"),
-            label = messages("messages__common__lastname")
-        )
-
-        @components.input_date(
-            field = form("date"),
-            label = messages("messages__common__dob"),
-            hint = Some(messages("messages__common__dob__hint"))
-        )
+        @components.person_details(form)
 
         @components.submit_button()
     }
